fix: revert some settings and add some error handling

+++ src/main/java/com/zoomers/GameSetMatch/services/TournamentService.java
@@ -2,16 +2,20 @@ package com.zoomers.GameSetMatch.services;
 
 import com.zoomers.GameSetMatch.entity.Tournament;
 import com.zoomers.GameSetMatch.repository.TournamentRepository;
+import com.zoomers.GameSetMatch.repository.UserRegistersTournamentRepository;
 import com.zoomers.GameSetMatch.scheduler.enumerations.TournamentStatus;
-import com.zoomers.GameSetMatch.services.Errors.EntityNotFoundError;
-import com.zoomers.GameSetMatch.services.Errors.LessThanRequiredNumOfRegistrantsError;
+import com.zoomers.GameSetMatch.services.Errors.InvalidActionForTournamentStatusException;
+import com.zoomers.GameSetMatch.services.Errors.MinRegistrantsNotMetException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.mail.MessagingException;
+import javax.persistence.EntityNotFoundException;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+
+import static java.util.Objects.isNull;
 
 @Service
 public class TournamentService {
@@ -47,12 +51,27 @@ public class TournamentService {
         return tournament.findTournaments(status, id);
     }
 
-    public void deleteTournamentByID(Integer id) throws MessagingException {
+    public void deleteTournamentByID(Integer id) throws MessagingException, InvalidActionForTournamentStatusException {
         // remove references in other tables
-        mailService.sendCancelMail(id);
+        Tournament t = this.findTournamentByID(id).orElse(null);
+
+        if (isNull(t)){
+            throw new EntityNotFoundException(String.format("Unable to find the tournament with id %d", id));
+        }
+
+        if (t.getStatus() >= TournamentStatus.ONGOING.getStatus()) {
+            throw new InvalidActionForTournamentStatusException("Cannot delete a tournament that is in progress");
+        }
+
+        List<UserRegistersTournamentRepository.IRegistrant> participants = userRegistersTournament.getRegistrants(id);
+        String tournamentName = t.getName();
+
         availability.deleteByTournamentID(id);
         userRegistersTournament.deleteByTournamentID(id);
         tournament.deleteTournamentByTournamentID(id);
+
+        mailService.sendCancelMail(participants, tournamentName);
+
     }
 
     public boolean changeTournamentStatus(Integer id, TournamentStatus status) {
@@ -68,14 +87,14 @@ public class TournamentService {
         return false;
     }
 
-    public void closeRegistration(Integer id) throws LessThanRequiredNumOfRegistrantsError, EntityNotFoundError {
+    public void closeRegistration(Integer id) throws MinRegistrantsNotMetException, EntityNotFoundException {
         Tournament tournament = this.findTournamentByID(id).orElse(null);
         Integer numRegistrants = userRegistersTournament.getNumberOfRegistrantsForATournament(id);
         if (tournament == null) {
-            throw new EntityNotFoundError(String.format("Unable to find the tournament with id %d", tournament.getTournamentID()));
+            throw new EntityNotFoundException(String.format("Unable to find the tournament with id %d", tournament.getTournamentID()));
 
         } else if (numRegistrants < tournament.getMinParticipants()) {
-            throw new LessThanRequiredNumOfRegistrantsError(
+            throw new MinRegistrantsNotMetException(
                     String.format("Minimum of %d players not reached. %d %s currently registered.",
                             tournament.getMinParticipants(),
                             numRegistrants,
